% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ou.R
\name{add_ou}
\alias{add_ou}
\title{Add Operating Unit Name}
\usage{
add_ou(df, operatingunit_name)
}
\arguments{
\item{df}{data frame from high frequency reporting}

\item{operatingunit_name}{OU name, quoted}
}
\description{
Add Operating Unit Name
}
\examples{
\dontrun{
#add missing OU name to reporting dataset
  df_uga <- add_ou(df_uga, "Uganda") }
}
