% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_agebands.R
\name{apply_agebands}
\alias{apply_agebands}
\title{Add mapped age columns onto dataframe}
\usage{
apply_agebands(df)
}
\arguments{
\item{df}{data frame with \code{agefine} or \code{agesemifine}}
}
\description{
Add mapped age columns onto dataframe
}
