% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_pds.R
\name{assign_pds}
\alias{assign_pds}
\title{Add HFR Period column}
\usage{
assign_pds(df, fy_start = 2019)
}
\arguments{
\item{df}{HFR data frame with date}

\item{fy_start}{start fiscal year, creates periods for 3 years, default = 2019}
}
\description{
Add HFR Period column
}
\examples{
\dontrun{
 df <- assign_pds(df) }
}
