% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_content}
\alias{check_content}
\title{Validate output content}
\usage{
check_content(df, output_path, datim_path)
}
\arguments{
\item{df}{HFR DataFrame}

\item{datim_path}{path of datim lookup files}
}
\value{
df updated HFR dataframe
}
\description{
Validate output content
}
