% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_disaggs}
\alias{check_disaggs}
\title{Validate disaggs for export}
\usage{
check_disaggs(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validate disaggs for export
}
