% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_import.R
\name{check_distinct_ous}
\alias{check_distinct_ous}
\title{Check OUs listed in operatingunit}
\usage{
check_distinct_ous(df)
}
\arguments{
\item{df}{df create during \code{hfr_import()}}
}
\description{
Check OUs listed in operatingunit
}
