% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_inds}
\alias{check_inds}
\title{Validate indicators for export}
\usage{
check_inds(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validate indicators for export
}
