% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_mechs}
\alias{check_mechs}
\title{Validate mechanisms for export}
\usage{
check_mechs(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validate mechanisms for export
}
