% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_orgunituids}
\alias{check_orgunituids}
\title{Validate orgunituids for export}
\usage{
check_orgunituids(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validate orgunituids for export
}
