% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{check_output_cols}
\alias{check_output_cols}
\title{Validate columns for export}
\usage{
check_output_cols(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validate columns for export
}
