% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_initial.R
\name{check_tabs}
\alias{check_tabs}
\title{Validate Submission's tabs for import}
\usage{
check_tabs(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Validate Submission's tabs for import
}
