% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_ken.R
\name{clean_vars_ken}
\alias{clean_vars_ken}
\title{Rename Variables to Unify Kenya Parnter HFR Submissions}
\usage{
clean_vars_ken(df)
}
\arguments{
\item{df}{Kenya HFR dataframe}
}
\description{
Rename Variables to Unify Kenya Parnter HFR Submissions
}
