% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sources.R
\name{combine_sources}
\alias{combine_sources}
\title{Joing HFR and DATIM Data}
\usage{
combine_sources(path_hfr, path_datim, start_date, hfr_pd, weeks = 4,
  output_folder = NULL)
}
\arguments{
\item{path_hfr}{file path to HFR processed data}

\item{path_datim}{file path to DATIM extract, see \code{extract_datim()}}

\item{start_date}{start date of HFR period}

\item{hfr_pd}{HFR period # (for saving)}

\item{weeks}{number of weeks to create, default = 4}

\item{output_folder}{folder path for saving the output}
}
\description{
Joing HFR and DATIM Data
}
\examples{
\dontrun{
#identify file paths
  path_datim <- list.files("out/DATIM", "FY19Q2_Zam", full.names = TRUE) \%>\% fs::path_abs()
  path_hfr <- list.files("out/processed", "ZMB", full.names = TRUE) \%>\% fs::path_abs()
#combine
  combine_sources(path_hfr, path_datim, "2019-06-10", 10, output_folder = "out/joint") }
}
