% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{confirm_validations}
\alias{confirm_validations}
\title{Confirm validations of processed files}
\usage{
confirm_validations(hfr_data, hfr_errors, dir_files)
}
\arguments{
\item{hfr_data}{content of processed files}

\item{hfr_errors}{errors detected from files content}

\item{dir_files}{location of processed files}
}
\description{
Confirm validations of processed files
}
\examples{
\dontrun{
confirm_validations(hfr_data, hfr_errors, dir_files)
}

}
