% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_missing}
\alias{count_missing}
\title{Count missing values}
\usage{
count_missing(df, var)
}
\arguments{
\item{df}{data frame}

\item{var}{variable to count missing values}
}
\description{
Count missing values
}
