% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{curr_fy}
\alias{curr_fy}
\title{Fiscal Year}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
curr_fy
}
\value{
Current Fiscal Year
}
\description{
Fiscal Year
}
\keyword{datasets}
