% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hfd.R
\name{export_hfd}
\alias{export_hfd}
\title{Export High Frequency Data}
\usage{
export_hfd(df, folderpath_output = NULL)
}
\arguments{
\item{df}{structured High Frequency Data Frame}

\item{folderpath_output}{provide the full path to the folder for saving}
}
\description{
Export High Frequency Data
}
\examples{
\dontrun{
 #write output
   export_hfd(df_tza, "TZA", "~/WeeklyData") }
}
