% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hfr.R
\name{export_hfr}
\alias{export_hfr}
\title{Export High Frequency Data}
\usage{
export_hfr(df, folderpath_output = NULL)
}
\arguments{
\item{df}{structured High Frequency Data Frame}

\item{folderpath_output}{provide the full path to the folder for saving}
}
\description{
Export High Frequency Data
}
\examples{
\dontrun{
 #write output
   export_hfr(df_tza, "TZA", "~/WeeklyData") }
}
