% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_datim_api.R
\name{extract_datim}
\alias{extract_datim}
\title{Extract DATIM Results and Targets (DATIM API Call)}
\usage{
extract_datim(mechanism_id = NULL, ou_name = NULL, username, password,
  baseurl = "https://final.datim.org/", quarters_complete = NULL,
  folderpath_output = NULL)
}
\arguments{
\item{mechanism_id}{DATIM mechanism ID}

\item{ou_name}{Operating Unit name, if mechanism is not specified}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}

\item{quarters_complete}{of quarters completed through FY to determine weeks left in year}

\item{folderpath_output}{folder path to store DATIM output, default = NULL}
}
\description{
Extract DATIM Results and Targets (DATIM API Call)
}
\examples{
\dontrun{
 #mechanism targets
 myuser <- "UserX"
 mech_x_targets <- extract_datim(mechanism_id = 00001, username = myuser, password = mypwd(myuser))
 #ou targets
 mech_x_targets <- extract_datim(ou_name = "Namibia", username = myuser, password = mypwd(myuser))
 }
}
