% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_headers.R
\name{extract_headers}
\alias{extract_headers}
\title{Extract Headers}
\usage{
extract_headers(filepath, sheetname, skip_lines)
}
\arguments{
\item{filepath}{full filepath to the country file}

\item{sheetname}{tab name}

\item{skip_lines}{number of lines to skip}
}
\description{
Extract Headers
}
\examples{
#' \dontrun{
#get new header row
  path <- "~/WeeklyData/Cameroon_data.xlsx"
  headers <- extract_headers(path, "TX_NEW", 1) }
}
