% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{extract_iso3code}
\alias{extract_iso3code}
\title{Extract OU ISO3 code}
\usage{
extract_iso3code(pfile)
}
\arguments{
\item{pfile}{processed file}
}
\value{
ISO3 3 character iso code
}
\description{
Extract OU ISO3 code
}
\examples{
\dontrun{
  extract_iso3code('HFR_2020.99_XAR_100000_processed_20200528.csv')
}

}
