% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{extract_mechcode}
\alias{extract_mechcode}
\title{Extract mechanism code}
\usage{
extract_mechcode(pfile)
}
\arguments{
\item{pfile}{processed file}
}
\value{
mech_code mechanism code
}
\description{
Extract mechanism code
}
\examples{
\dontrun{
  extract_mechcode('HFR_2020.99_XAR_100000_processed_20200528.csv')
}

}
