% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_api.R
\name{extract_targets}
\alias{extract_targets}
\title{Extract Targets (DATIM API Call)}
\usage{
extract_targets(mechanism_id, username, password,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{mechanism_id}{DATIM mechanism ID}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Extract Targets (DATIM API Call)
}
\examples{
\dontrun{
 myuser <- "UserX"
 mech_x_targets <- extract_targets(00001, myuser, mypwd(myuser))}
}
