% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_date.R
\name{fix_date}
\alias{fix_date}
\title{Convert dates to date format}
\usage{
fix_date(df, round_hfrdate = FALSE)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}

\item{round_hfrdate}{rounds date to the nearest HFRweek start (for non-compliance), default = FALSE}
}
\description{
Convert dates to date format
}
