% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_ken.R
\name{fix_dates_ken}
\alias{fix_dates_ken}
\title{Conform dates from different partners}
\usage{
fix_dates_ken(df, filepath)
}
\arguments{
\item{df}{Kenya HFR data frame}

\item{filepath}{Kenya HFR file path}
}
\description{
Conform dates from different partners
}
