% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_noncompliance.R
\name{fix_noncompliance}
\alias{fix_noncompliance}
\title{Resolve issues with non-standard entries}
\usage{
fix_noncompliance(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}
}
\description{
Resolve issues with non-standard entries
}
