% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flag_extra}
\alias{flag_extra}
\title{Flag Extra Variables}
\usage{
flag_extra(required, submitted)
}
\arguments{
\item{required}{list of required vars}

\item{submitted}{list of vars pulled from submission}
}
\description{
Flag Extra Variables
}
