% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{flag_missing}
\alias{flag_missing}
\title{Flag Missing Variables}
\usage{
flag_missing(required, submitted)
}
\arguments{
\item{required}{list of required vars}

\item{submitted}{list of vars pulled from submission}
}
\description{
Flag Missing Variables
}
