% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_hfr.R
\name{gather_hfr}
\alias{gather_hfr}
\title{Reshape HFR Data frame long}
\usage{
gather_hfr(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}
}
\description{
Reshape HFR Data frame long
}
