% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_hts_num.R
\name{gen_hts_num}
\alias{gen_hts_num}
\title{Gernate HTS_TST Total Numerator from POS and NEG}
\usage{
gen_hts_num(df)
}
\arguments{
\item{df}{HFR processed dataframe to create a HTS total numerator}
}
\description{
Gernate HTS_TST Total Numerator from POS and NEG
}
