% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mer.R
\name{gen_url}
\alias{gen_url}
\title{Generate a API URL}
\usage{
gen_url(
  ou_uid,
  org_lvl,
  org_type = "facility",
  type_hts = NULL,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ou_uid}{UID for the country, recommend using \code{identify_ouuids()}}

\item{org_lvl}{org hierarchy level, eg facility is level 7 in country X, recommend using \code{identify_levels()}}

\item{org_type}{organization type, either facility (default) or community}

\item{type_hts}{is the API call for HTS indicators ("results", "targets"), default = NULL}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Generate a API URL
}
\examples{
\dontrun{
 #get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Ghana")
 #get facility level
  faclvl <- identify_levels("Ghana", "facility", username = myuser, password = mypwd())
 #gen url
  myurl <- gen_url(ouuid, faclvl, org_type = facility, type_hts = NULL) }
}
