% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{get_operatingunit}
\alias{get_operatingunit}
\title{Get operating unit name}
\usage{
get_operatingunit(iso_code, orglevels)
}
\arguments{
\item{iso_code}{iso3 code}

\item{orglevels}{df org levels}
}
\value{
operating unit
}
\description{
Get operating unit name
}
\examples{
\dontrun{
  get_operatingunit(org_levels, 'XWA')
}

}
