% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{guess_operatingunit}
\alias{guess_operatingunit}
\title{Get operating unit name}
\usage{
guess_operatingunit(levels, ims, pfile)
}
\arguments{
\item{levels}{org levels}

\item{ims}{mechanisms df}

\item{pfile}{processed file}
}
\value{
operating unit name
}
\description{
Get operating unit name
}
