% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_append_sources.R
\name{hfr_append_sources}
\alias{hfr_append_sources}
\title{#' Append HFR and DATIM Data}
\usage{
hfr_append_sources(
  folderpath_hfr,
  folderpath_datim,
  start_date,
  weeks = 4,
  max_date = TRUE,
  folderpath_output
)
}
\arguments{
\item{folderpath_hfr}{folder path to HFR processed data}

\item{folderpath_datim}{folder path to DATIM extracts, see \code{extract_datim()}}

\item{start_date}{start date of HFR period, YYYY-MM-DD format}

\item{weeks}{number of weeks to create, default = 4}

\item{max_date}{cut off data at max date? default = NULL}

\item{folderpath_output}{folder path for saving the output}
}
\description{
#' Append HFR and DATIM Data
}
