% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_assign_pds.R
\name{hfr_assign_pds}
\alias{hfr_assign_pds}
\title{Add HFR Period column}
\usage{
hfr_assign_pds(df)
}
\arguments{
\item{df}{HFR data frame with date}
}
\description{
Add HFR Period column
}
\examples{
\dontrun{
 df <- hfr_assign_pds(df) }
}
