% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_batch_read.R
\name{hfr_batch_read}
\alias{hfr_batch_read}
\title{Batch read hfr files}
\usage{
hfr_batch_read(pfolder, pattern, source = FALSE)
}
\arguments{
\item{pfolder}{processed file folder}

\item{pattern}{filename pattern}

\item{source}{append source filename?}
}
\value{
df
}
\description{
Batch read hfr files
}
