% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_export.R
\name{hfr_export}
\alias{hfr_export}
\title{Export High Frequency Data}
\usage{
hfr_export(df, folderpath_output = NULL, quarters_complete = NULL)
}
\arguments{
\item{df}{structured High Frequency Data Frame}

\item{folderpath_output}{provide the full path to the folder for saving}

\item{quarters_complete}{FOR DATIM ONLY: # of quarters completed through FY to determine weeks left in year}
}
\description{
Export High Frequency Data
}
\examples{
\dontrun{
 #write output
   hfr_export(df_tza, "~/WeeklyData") }
}
