% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_export.R
\name{hfr_export_mech}
\alias{hfr_export_mech}
\title{Export csv files by mechanism}
\usage{
hfr_export_mech(df, mech, type, folderpath_output, quarters_complete)
}
\arguments{
\item{df}{tructured High Frequency Data Frame}

\item{mech}{mech_code}

\item{type}{type type of data being saved, default = processed}

\item{folderpath_output}{provide the full path to the folder for saving}

\item{quarters_complete}{FOR DATIM ONLY: # of quarters completed through FY to determine weeks left in year}
}
\description{
Export csv files by mechanism
}
