% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hfr_extract_meta}
\alias{hfr_extract_meta}
\title{Extract Meta Data Information about Template}
\usage{
hfr_extract_meta(filepath, meta_type = "type")
}
\arguments{
\item{filepath}{filepath to sumbitted template}

\item{meta_type}{type of meta data requesting: ou, period, version, type (default)}
}
\description{
Useful for pulling information about the template, whether
It be the Operating Unit (OU), Period, template version, or type, eg wide or long.
}
\examples{
\dontrun{
#identify whether template is long or wide
  filepath <- "~/WeeklyData/Raw/KEN_Weekly.xlsx"
  hfr_extract_meta(filepath, meta_type = "type")
#identify period
  hfr_extract_meta(filepath, meta_type = "period")
#identify OU
  hfr_extract_meta(filepath, meta_type = "ou") }
}
