% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_fix_date.R
\name{hfr_fix_date}
\alias{hfr_fix_date}
\title{Convert dates to date format}
\usage{
hfr_fix_date(df, round_hfrdate = FALSE)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}

\item{round_hfrdate}{rounds date to the nearest HFRweek start (for non-compliance), default = FALSE}
}
\description{
Convert dates to date format
}
