% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_fix_noncompliance.R
\name{hfr_fix_noncompliance}
\alias{hfr_fix_noncompliance}
\title{Resolve issues with non-standard entries}
\usage{
hfr_fix_noncompliance(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}
}
\description{
Resolve issues with non-standard entries
}
