% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_gap_target.R
\name{hfr_gap_target}
\alias{hfr_gap_target}
\title{Generate Gap Targets}
\usage{
hfr_gap_target(df, quarters_complete)
}
\arguments{
\item{df}{data frame created by \code{extract_datim()}}

\item{quarters_complete}{MER quarters with data available}
}
\description{
Generate Gap Targets
}
\examples{
\dontrun{
 myuser <- "UserX"
 mech_x_targets <- extract_datim(00001, myuser, mypwd(myuser))
 mech_x_targets <- hfr_gap_target(mech_x_targets, 2) }
}
