% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_gather.R
\name{hfr_gather}
\alias{hfr_gather}
\title{Reshape HFR Data frame long}
\usage{
hfr_gather(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}
}
\description{
Reshape HFR Data frame long
}
