% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_aggr.R
\name{hfr_group_wkly}
\alias{hfr_group_wkly}
\title{Create Monthly Aggregate for Weekly data}
\usage{
hfr_group_wkly(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}
}
\description{
Create Monthly Aggregate for Weekly data
}
