% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_fix_date.R
\name{hfr_identify_freq}
\alias{hfr_identify_freq}
\title{Identify reporting frequency}
\usage{
hfr_identify_freq(df)
}
\arguments{
\item{df}{HFR data frame from \code{hfr_fix_date()}}
}
\description{
Identify reporting frequency
}
