% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_assign_pds.R
\name{hfr_identify_pds}
\alias{hfr_identify_pds}
\title{Create a data frame of HFR weeks and periods}
\usage{
hfr_identify_pds(fy = NULL)
}
\arguments{
\item{fy}{fiscal year}
}
\description{
Create a data frame of HFR weeks and periods
}
