% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_import.R
\name{hfr_import}
\alias{hfr_import}
\title{Import template sheet(s)}
\usage{
hfr_import(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Import template sheet(s)
}
\examples{
\dontrun{
#identify whether template is long or wide
  filepath <- "~/WeeklyData/Raw/KEN_Weekly.xlsx"
  df_hfr <- hfr_import(filepath) }
}
