% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_munge_string.R
\name{hfr_munge_string}
\alias{hfr_munge_string}
\title{clean up/standardize string text for indicators and disaggs}
\usage{
hfr_munge_string(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}
}
\description{
clean up/standardize string text for indicators and disaggs
}
