% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hfr_orgunit_search}
\alias{hfr_orgunit_search}
\title{Search Org Hierarchy for Org Unit}
\usage{
hfr_orgunit_search(df, orgunit_name, ou = NULL)
}
\arguments{
\item{df}{org hierarchy, created in pull_hierarchy()}

\item{orgunit_name}{full or partial orgunit name for matching}

\item{ou}{operating unit; if added searches only that OU default = NULL}
}
\description{
Search Org Hierarchy for Org Unit
}
\examples{
\dontrun{
# orgunit - Kewot
org <- pull_hierarchy(myusername, mypass)
hfr_orgunit_search(org, "Kew", "Ethiopia") }
}
