% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_process_template.R
\name{hfr_process_template}
\alias{hfr_process_template}
\title{Import and Munge HFR Standard Templates}
\usage{
hfr_process_template(filepath, round_hfrdate = FALSE,
  folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to sumbitted template}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}

\item{round_date}{rounds date to the nearest HFRweek start (for non-compliance), default = FALSE}
}
\description{
Import and Munge HFR Standard Templates
}
