% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_read.R
\name{hfr_read}
\alias{hfr_read}
\title{Read in HFR output file}
\usage{
hfr_read(filepath)
}
\arguments{
\item{filepath}{filepath of an HFR output file}
}
\description{
Read in HFR output file
}
\examples{
\dontrun{
 path <- "~/data/HFR_2020.01_Global_output_20191204.1705.csv"
 df <- hfr_read(path)
 }
}
