% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_read_all.R
\name{hfr_read_all}
\alias{hfr_read_all}
\title{Batch read hfr files}
\usage{
hfr_read_all(pfolder, pattern, source = FALSE)
}
\arguments{
\item{pfolder}{processed file folder}

\item{pattern}{filename pattern}

\item{source}{append source filename?}
}
\value{
df
}
\description{
Batch read hfr files
}
\examples{
\dontrun{
  hfr_read_all('./Data/2020.05', pattern='HFR_FY2020.05')
}

}
