% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_restrict_cols.R
\name{hfr_restrict_cols}
\alias{hfr_restrict_cols}
\title{Restrict HFR data frame columns}
\usage{
hfr_restrict_cols(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{hfr_import()}}
}
\description{
Restrict HFR data frame columns
}
