% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_fix_date.R
\name{hfr_round_date}
\alias{hfr_round_date}
\title{Round to nearest HFR date}
\usage{
hfr_round_date(df)
}
\arguments{
\item{df}{df HFR data frame imported via \code{hfr_import()}}
}
\description{
Round to nearest HFR date
}
