% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_validation.R
\name{hfr_validation}
\alias{hfr_validation}
\title{Validation Checks}
\usage{
hfr_validation(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validation Checks
}
