% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_pull.R
\name{hierarchy_clean}
\alias{hierarchy_clean}
\title{Clean up DATIM Hierarchy Path}
\usage{
hierarchy_clean(df)
}
\arguments{
\item{df}{data frame created by \code{hierarchy_pull()}}
}
\description{
Clean up DATIM Hierarchy Path
}
