% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{hierarchy_extract}
\alias{hierarchy_extract}
\title{Pull Hierarchy Data from DATIM}
\usage{
hierarchy_extract(
  ou_uid,
  username,
  password,
  baseurl = "https://final.datim.org/"
)
}
\arguments{
\item{ou_uid}{UID for the country, recommend using \code{identify_ouuids()}}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Pull Hierarchy Data from DATIM
}
\examples{
\dontrun{
#get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Kenya")
#pull hierarchy (paths are all UIDs)
  df <- hierarchy_extract(ouuid, username = myuser, password = mypwd(myuser)) }
}
