% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_pull.R
\name{hierarchy_identify_ctry}
\alias{hierarchy_identify_ctry}
\title{Extract country name from OU or country name}
\usage{
hierarchy_identify_ctry(df)
}
\arguments{
\item{df}{data frame created by \code{hierarchy_pull() \%>\% hierarchy_clean()}}
}
\description{
Extract country name from OU or country name
}
