% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy_pull.R
\name{hierarchy_pull}
\alias{hierarchy_pull}
\title{Pull Hierarchy Data from DATIM}
\usage{
hierarchy_pull(ou_uid, username, password,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{ou_uid}{UID for the country, recommend using \code{identify_ouuids()}}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Pull Hierarchy Data from DATIM
}
\examples{
\dontrun{
#get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Kenya")
#pull hierarchy (paths are all UIDs)
  df <- hierarchy_pull(ouuid, username = myuser, password = mypwd(myuser)) }
}
