% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{hierarchy_rename}
\alias{hierarchy_rename}
\title{Rename Hierarchy from Levels to OU/SNU1/PSNU/Facility}
\usage{
hierarchy_rename(df, country, username, password,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{df}{data frame created by \code{hierarchy_extract() \%>\% hierarchy_clean()}}

\item{country}{county name, eg "Malawi" or "Nepal"}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Rename Hierarchy from Levels to OU/SNU1/PSNU/Facility
}
