% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_headers_bdi.R
\name{identify_headers_bdi}
\alias{identify_headers_bdi}
\title{Create un-umerged, unique header row}
\usage{
identify_headers_bdi(filepath)
}
\arguments{
\item{filepath}{full filepath to the country file}
}
\description{
Create un-umerged, unique header row
}
\examples{
\dontrun{
#get new header row
  path <- "~/WeeklyData/Burundi_data.xlsx"
  headers <- identify_headers_bdi(path)
#apply new headers to the import
 df_bdi <- readxl::read_excel(path, col_names = headers) }
}
