% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_headers_cmr.R
\name{identify_headers_cmr}
\alias{identify_headers_cmr}
\title{Create un-umerged, unique header row for Cameroon}
\usage{
identify_headers_cmr(filepath, sheetname)
}
\arguments{
\item{filepath}{full filepath to the country file}

\item{sheetname}{tab name}
}
\description{
Create un-umerged, unique header row for Cameroon
}
\examples{
\dontrun{
#get new header row
  path <- "~/WeeklyData/Cameroon_data.xlsx"
  headers <- identify_headers_cameroon(path, "TX_NEW")
#apply new headers to the import
 df_hti <- readxl::read_excel(path, sheet = "TX_NEW", skip = 1, col_names = headers) }
}
