% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_headers_hti.R
\name{identify_headers_hti}
\alias{identify_headers_hti}
\title{Create un-umerged, unique header row}
\usage{
identify_headers_hti(filepath, sheetname)
}
\arguments{
\item{filepath}{full filepath to the country file}

\item{sheetname}{tab name}
}
\description{
Create un-umerged, unique header row
}
\examples{
\dontrun{
#get new header row
  path <- "~/WeeklyData/Haiti_data.xlsx"
  headers <- identify_headers_hti(path, "ALL")
#apply new headers to the import
 df_hti <- readxl::read_excel(path, sheet = "ALL", skip = 4, col_names = headers) }
}
