% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_datim_api.R
\name{identify_levels}
\alias{identify_levels}
\title{Identify Facility/Community levels in org hierarchy}
\usage{
identify_levels(ou = NULL, username, password,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{ou}{operating unit name}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{base API url, default = https://final.datim.org/}
}
\description{
Identify Facility/Community levels in org hierarchy
}
\examples{
\dontrun{
 #table for all OUs
  myuser <- "UserX"
  identify_levels(username = myuser, password = mypwd())
 #table for just Kenya
   identify_levels("Kenya", username = myuser, password = mypwd())
 #facility level of Kenya
   identify_levels("Kenya", "facility", username = myuser, password = mypwd()) }
}
