% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_datim_api.R
\name{identify_mechs}
\alias{identify_mechs}
\title{Identify active mechanisms by OU}
\usage{
identify_mechs(curr_fy = 2020, mechid = NULL, extract = NULL,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{curr_fy}{current fiscal year to filter by, default = 2020}

\item{mechid}{mechanism id to filter by, optional}

\item{extract}{variable to extract from the mechanism table, optional}

\item{baseurl}{API base url}
}
\description{
Identify active mechanisms by OU
}
\examples{
\dontrun{
 #full active mechanism table
  identify_mechs()
 #mechanism information for a particular mechanism
  identify_mechs(mechid = 00001)
 #uid for a particular mechanism
  identify_mechs(mechid = 00001, extract = "uid") }
}
