% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_pds.R
\name{identify_pds}
\alias{identify_pds}
\title{Create a data frame of HFR weeks and periods}
\usage{
identify_pds(fy = NULL)
}
\arguments{
\item{fy}{fiscal year}
}
\description{
Create a data frame of HFR weeks and periods
}
