% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_hfr.R
\name{import_hfr}
\alias{import_hfr}
\title{Import template sheet(s)}
\usage{
import_hfr(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Import template sheet(s)
}
\examples{
\dontrun{
#identify whether template is long or wide
  filepath <- "~/WeeklyData/Raw/KEN_Weekly.xlsx"
  df_hfr <- (filepath) }
}
