% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_mwi.R
\name{import_mwi_comm}
\alias{import_mwi_comm}
\title{Import disperate community sheets}
\usage{
import_mwi_comm(filepath, sht)
}
\arguments{
\item{filepath}{Malawi HFR Community excel filepath}

\item{sht}{sheet name}
}
\description{
Import disperate community sheets
}
