% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_map_bdi-data.R
\docType{data}
\name{ind_map_bdi}
\alias{ind_map_bdi}
\title{Burundi Indicator Mapping Table}
\format{A data frame including all relevant indicators
\describe{
\item{ind}{indicator & disaggregate name used by BDI}
\item{indicator}{standardized indicator to work across partners and OUs}
\item{disaggregate}{components of the indicator that take direction from the MER}
\item{otherdisaggregate}{additional components of the indicator, namely days for SDI}
}}
\usage{
data(ind_map_bdi)
}
\description{
Table that cross references between partner data
collected/reported to the mission monthly
and the PEPFAR MER indicators.
}
\keyword{datasets}
