% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_map_lso-data.R
\docType{data}
\name{ind_map_lso}
\alias{ind_map_lso}
\title{Lesotho Indicator Mapping Table}
\format{A data frame including all relevant indicators
\describe{
\item{ind}{indicator reported by IM}
\item{indicator}{standardized indicator to work across partners and OUs}
\item{disaggregate}{components of the indicator that take direction from the MSD}
\item{sex}{sex, male or female}
\item{agecoarse}{age, <15/15+}
\item{resultstatus}{HIV status, Positive/Negative}
\item{otherdisaggregate}{other components of the indicator not captured elsewhere}
}}
\usage{
data(ind_map_lso)
}
\description{
Table that cross references between partner data
collected/reported to the mission weekly
and the PEPFAR MER indicators.
}
\keyword{datasets}
