% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_map_uga-data.R
\docType{data}
\name{ind_map_uga}
\alias{ind_map_uga}
\title{Uganda Indicator Mapping Table}
\format{A data frame including all relevant indicators
\describe{
\item{indicator_orig}{indicator name used by UGA}
\item{disaggregate_orig}{disaggregate name used by UGA}
\item{indicator}{standardized indicator to work across partners and OUs}
\item{disaggregate}{components of the indicator that take direction from the MER}
\item{agecoarse}{age, <15/15+}
\item{agesemifine}{VMMC age bands, <15, 15-29, 30+}
\item{sex}{sex, male, female, or peds}
\item{modality}{testing modalities (MER)}
}}
\usage{
data(ind_map_uga)
}
\description{
Table that cross references between partner data
collected/reported to the mission monthly
and the PEPFAR MER indicators.
}
\keyword{datasets}
