% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_checks.R
\name{is_date_valid}
\alias{is_date_valid}
\title{Validate reporting dates}
\usage{
is_date_valid(.data, df_dates)
}
\arguments{
\item{.data}{df}

\item{df_dates}{df_dates}
}
\value{
boolean
}
\description{
Validate reporting dates
}
\examples{
\dontrun{
  data \%>\% is_date_valid(df_dates)
}

}
