% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_initial.R
\name{is_hfrtab}
\alias{is_hfrtab}
\title{Determine if there are tabs to import}
\usage{
is_hfrtab(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Determine if there are tabs to import
}
