% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_date.R
\name{is_iso_date}
\alias{is_iso_date}
\title{Determine if dates are in Excel format}
\usage{
is_iso_date(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}
}
\description{
Determine if dates are in Excel format
}
