% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_checks.R
\name{is_mech_valid}
\alias{is_mech_valid}
\title{Validate mechanism code}
\usage{
is_mech_valid(.data, df_mechs)
}
\arguments{
\item{.data}{dataset}

\item{df_mechs}{mechs df}
}
\value{
boolean
}
\description{
Validate mechanism code
}
