% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_metatab}
\alias{is_metatab}
\title{Determine whether meta tab exists}
\usage{
is_metatab(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Determine whether meta tab exists
}
