% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_checks.R
\name{is_ou_valid}
\alias{is_ou_valid}
\title{Check if OperatingUnit is valid}
\usage{
is_ou_valid(.data, df_orgs)
}
\arguments{
\item{data}{dataset}

\item{df_dates}{dates df}
}
\value{
boolean
}
\description{
Check if OperatingUnit is valid
}
