% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_map-data.R
\docType{data}
\name{iso_map}
\alias{iso_map}
\title{ISO Code Mapping Table}
\format{A data frame with each PEPFAR OU and associated ISO code.
\describe{
\item{operatingunit}{PEPFAR Operating Unit or country}
\item{iso}{3 letter ISO code for the Operatingunit or country}
\item{regional}{TRUE if operatingunit is a country under a regional program}
}}
\source{
\url{"https://final.datim.org/api/dataStore/dataSetAssignments/orgUnitLevels"}
}
\usage{
data(iso_map)
}
\description{
Table that cross references the ISO Country codes to the
PEPFAR Operating Units and Countries. This table is used
during the export process for file naming. ISO codes were
originall pulled from DATIM.
}
\keyword{datasets}
