% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lookups.R
\name{load_lookups}
\alias{load_lookups}
\title{Load DATIM Look up tables}
\usage{
load_lookups(datim_path = "../datim")
}
\arguments{
\item{datim_path}{location of datim look up tables files}
}
\description{
Load DATIM Look up tables
}
\examples{
\dontrun{
 #load look up data
 load_lookups(datim_path = "./Data/datim")}
}
