% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge_string.R
\name{munge_string}
\alias{munge_string}
\title{clean up/standardize string text for indicators and disaggs}
\usage{
munge_string(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}
}
\description{
clean up/standardize string text for indicators and disaggs
}
