% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mypwd}
\alias{mypwd}
\title{Proper credentials from secure file}
\usage{
mypwd(username)
}
\arguments{
\item{username}{DATIM username}
}
\value{
see keyringr vignette about setting up storing secure credentials
}
\description{
Proper credentials from secure file
}
\examples{
\dontrun{
  myuser <- "UserX"
  mypwd(myuser) }
}
