% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{parse_submission}
\alias{parse_submission}
\title{Parse out submitted file components}
\usage{
parse_submission(pfile)
}
\arguments{
\item{sfile}{processed file}
}
\value{
component As vector c("hfr_pd", "iso3", "mech_code", "pdate")
}
\description{
Parse out submitted file components
}
\examples{
\dontrun{
  parse_submission("HFR_2020.99_XWH_100000_processed_20200101.csv")
}

}
