% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_hierarchy.R
\name{pull_hierarchy}
\alias{pull_hierarchy}
\title{Compile PEPFAR Hierarchy}
\usage{
pull_hierarchy(
  ou_uid,
  username,
  password,
  baseurl = "https://final.datim.org/",
  folderpath_output = NULL
)
}
\arguments{
\item{ou_uid}{UID for the country, recommend using \code{identify_ouuids()}}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}

\item{folderpath_output}{provide the full path to the folder for saving}
}
\description{
Compile PEPFAR Hierarchy
}
\examples{
\dontrun{
#get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Kenya")
#pull hierarchy (paths are all UIDs)
  df <- pull_hierarchy(ouuid, username = myuser, password = mypwd(myuser)) }
}
