% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{report_submissions_errors}
\alias{report_submissions_errors}
\title{Report files validation}
\usage{
report_submissions_errors(mechanisms, df_files, export = FALSE)
}
\arguments{
\item{mechanisms}{mechs df}

\item{df_files}{df of filename validation}
}
\value{
void
}
\description{
Report files validation
}
