% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_long.R
\name{reshape_long}
\alias{reshape_long}
\title{Reshape data long}
\usage{
reshape_long(df, meta_group)
}
\arguments{
\item{df}{wide data frame to reshape}

\item{meta_group}{meta data, no values, eg PSNU, partner, indicator}
}
\description{
Reshape data long
}
