% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{revert_validations}
\alias{revert_validations}
\title{Revert validated file to processed files}
\usage{
revert_validations(dir_files)
}
\arguments{
\item{dir_files}{location of processed files}
}
\description{
Revert validated file to processed files
}
\examples{
\dontrun{
revert_validations(dir_files)
}

}
