% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_date.R
\name{round_date}
\alias{round_date}
\title{Round to nearest HFR date}
\usage{
round_date(df)
}
\arguments{
\item{df}{df HFR data frame imported via \code{import_hfr()}}
}
\description{
Round to nearest HFR date
}
