% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_bdi.R
\name{structure_bdi}
\alias{structure_bdi}
\title{Import and Structure Burundi}
\usage{
structure_bdi(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the weekly Burundi
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Burundi's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for Burundi
  path <- "~/WeeklyData/Raw/BDI_monthly_data.xlsx"
  structure_hti(path)
#structure output for Burundi & export to txt file
  structure_hti(path, "~/WeeklyData/Output") }
}
