% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_civ.R
\name{structure_civ}
\alias{structure_civ}
\title{Import and Structure Cote d'Ivoire}
\usage{
structure_civ(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Cote d'Ivoire HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Cote d'Ivoire's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for Cote d'Ivoire
  path <- "~/WeeklyData/Raw/CIV_monthly_data.xlsx"
  structure_civ(path)
#structure output for Cote d'Ivoire & export to txt file
  structure_civ(path, "~/WeeklyData/Output") }
}
