% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_cod.R
\name{structure_cod}
\alias{structure_cod}
\title{Import and Structure DRC}
\usage{
structure_cod(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the monthly DRC
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from DRC's partner's monthly
reporting template files.
}
\examples{
\dontrun{
#structure output for DRC
  path <- "~/WeeklyData/Raw/DRC_montly_partner_x.xlsx"
  structure_cod(path)
#structure output for DRC & export to txt file
  structure_cod(path, "~/WeeklyData/Output") }
}
