% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_eth.R
\name{structure_eth}
\alias{structure_eth}
\title{Import and Structure Ethiopia}
\usage{
structure_eth(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Ethiopia HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Ethiopia
  path <- "~/WeeklyData/Raw/ETH_Weekly.xlsx"
  structure_eth(path)
#structure output for Ethiopia & export to txt file
  structure_eth(path, "~/WeeklyData/Output") }
}
