% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_hti.R
\name{structure_hti}
\alias{structure_hti}
\title{Import and Structure Haiti}
\usage{
structure_hti(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the monthly Haiti
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Haiti's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for Hait
  path <- "~/WeeklyData/Raw/HTI_monthly_data.xlsx"
  structure_hti(path)
#structure output for Haiti & export to txt file
  structure_hti(path, "~/WeeklyData/Output") }
}
