% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_ken.R
\name{structure_ken}
\alias{structure_ken}
\title{Import and Structure Kenya}
\usage{
structure_ken(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Kenya HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Kenya
  path <- "~/WeeklyData/Raw/KEN_Weekly.xlsx"
  structure_ken(path)
#structure output for Kenya & export to txt file
  structure_ken(path, "~/WeeklyData/Output") }
}
