% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_lso.R
\name{structure_lso}
\alias{structure_lso}
\title{Import and Structure Lesotho}
\usage{
structure_lso(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the weekly Lesotho
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Lesotho
  path <- "~/WeeklyData/Raw/LSO_Weekly.xlsx"
  structure_lso(path)
#structure output for Lesotho & export to txt file
  structure_lso(path, "~/WeeklyData/Output") }
}
