% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_moz.R
\name{structure_moz}
\alias{structure_moz}
\title{Import and Structure Mozambique}
\usage{
structure_moz(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Mozambique HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Mozambique
  path <- "~/WeeklyData/Raw/MOZ_Weekly.xlsx"
  structure_moz(path)
#structure output for Mozambique & export to txt file
  structure_moz(path, "~/WeeklyData/Output") }
}
