% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_mwi.R
\name{structure_mwi}
\alias{structure_mwi}
\title{Import and Structure Malawi}
\usage{
structure_mwi(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Malawi HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Malawi
  path <- "~/WeeklyData/Raw/MWI_Weekly.xlsx"
  structure_eth(path)
#structure output for Malawi & export to txt file
  structure_eth(path, "~/WeeklyData/Output") }
}
