% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_tza.R
\name{structure_tza}
\alias{structure_tza}
\title{Import and Structure Tanzania}
\usage{
structure_tza(filepath, type = "weekly", folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{type}{is the dataset a weekly or monthly file, "weekly", "monthly"}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the monthly Tanzania
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Tazania's monthly reporting
platform \url{http://hmis.reachproject.or.tz/MonthlyPortal/}
}
\examples{
\dontrun{
#structure output for Tanzania
  path <- "~/WeeklyData/Raw/All Monthly Raw Data in given timeframe.xlsx"
  structure_tza(path)
#structure output for Tanzania & export to txt file
  structure_tza(path, "~/WeeklyData/Output") }
}
