% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_uga.R
\name{structure_uga}
\alias{structure_uga}
\title{Import and Structure South Africa}
\usage{
structure_uga(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the weekly Uganda
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Uganda's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for UGA
  path <- "~/WeeklyData/Raw/UGA_partner_x.xlsx"
  structure_cod(path)
#structure output for UGA & export to txt file
  structure_cod(path, "~/WeeklyData/Output") }
}
