% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_ukr.R
\name{structure_ukr}
\alias{structure_ukr}
\title{Import and Structure Ukraine}
\usage{
structure_ukr(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the weekly Ukraine
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Ukraine's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for UKR
  path <- "~/WeeklyData/Raw/UKR_partner_x.xlsx"
  structure_ukr(path)
#structure output for UKR & export to txt file
  structure_ukr(path, "~/WeeklyData/Output") }
}
