% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_zaf.R
\name{structure_zaf}
\alias{structure_zaf}
\title{Import and Structure South Africa}
\usage{
structure_zaf(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the daily/weekly SA
dataset and structure in a uniform way to be tidy and align with other
partners/OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from South Africa's partner's monthly
reporting template files.
}
\examples{
\dontrun{
#structure output for ZAF
  path <- "~/WeeklyData/Raw/ZAF_partner_x.xlsx"
  structure_cod(path)
#structure output for ZAF & export to txt file
  structure_cod(path, "~/WeeklyData/Output") }
}
