% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_zmb.R
\name{structure_zmb}
\alias{structure_zmb}
\title{Import and Structure Zambia}
\usage{
structure_zmb(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Zambia HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs.
}
\examples{
\dontrun{
#structure output for Zambia
  path <- "~/WeeklyData/Raw/ZMB_Weekly.xlsx"
  structure_zmb(path)
#structure output for Zambia & export to txt file
  structure_zmb(path, "~/WeeklyData/Output") }
}
