% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_zwe.R
\name{structure_zwe}
\alias{structure_zwe}
\title{Import and Structure Zimbabwe}
\usage{
structure_zwe(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
This function will be used to take the Zimbabwe HFR
dataset and structure in a uniform way to be tidy and align with other
OU datasets. Only selected indicators will be imported and transformed
based on OHA's needs. The data is sourced from Zimbabwe's partner weekly
reporting template files.
}
\examples{
\dontrun{
#structure output for Zimbabwe
  path <- "~/WeeklyData/Raw/zwe_monthly_data.xlsx"
  structure_zwe(path)
#structure output for Zimbabwe & export to txt file
  structure_zwe(path, "~/WeeklyData/Output") }

}
