% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cols.R
\docType{data}
\name{template_cols_long}
\alias{template_cols_long}
\title{Column Headers for HFR Long Template}
\format{A list of all headers
\describe{
\item{template_cols_long}{column names}
}}
\usage{
data(template_cols_long)
}
\description{
List of column headers for the HFR Long Template
}
\keyword{datasets}
