% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cols.R
\docType{data}
\name{template_cols_meta}
\alias{template_cols_meta}
\title{Column Headers for HFR Indicator Meta Data}
\format{A list of all headers
\describe{
\item{template_cols_meta}{column names}
}}
\usage{
data(template_cols_meta)
}
\description{
List of column headers for the HFR Indicator Meta Data
}
\keyword{datasets}
