% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cols.R
\docType{data}
\name{template_cols_wide}
\alias{template_cols_wide}
\title{Column Headers for HFR Wide Template}
\format{A list of all headers
\describe{
\item{template_cols_wide}{column names}
}}
\usage{
data(template_cols_wide)
}
\description{
List of column headers for the HFR Wide Template
}
\keyword{datasets}
