% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_cols.R
\docType{data}
\name{template_cols_wide_lim}
\alias{template_cols_wide_lim}
\title{Column Headers for HFR Wide - Limited Template}
\format{A list of all headers
\describe{
\item{template_cols_wide_lim}{column names}
}}
\usage{
data(template_cols_wide_lim)
}
\description{
List of column headers for the HFR Wide Template
}
\keyword{datasets}
