% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_sitelist.R
\name{tidy_sitelist}
\alias{tidy_sitelist}
\title{Title}
\usage{
tidy_sitelist(filepath, folderpath_output = NULL)
}
\arguments{
\item{filepath}{path to sitelist file}

\item{folderpath_output}{if output is desired, full folder path}
}
\description{
Title
}
\examples{
\dontrun{
  path <- "~/Data/HFR_FY21_SiteValidation_Kenya.xlsx"
  df_sites <- tidy_sitelist(path) }
}
