% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unify_hfr.R
\name{unify_hfr}
\alias{unify_hfr}
\title{Unify High Frequency Reporting Dataset}
\usage{
unify_hfr(ou_iso, filepath, folderpath_output = NULL)
}
\arguments{
\item{ou_iso}{ISO code for the operating unit}

\item{filepath}{filepath to the output file}

\item{folderpath_output}{if a txt output is desired, provide the full path to the folder}
}
\description{
Unify High Frequency Reporting Dataset
}
\examples{
\dontrun{
#structure output for Tanzania
  path <- "~/WeeklyData/Raw/All Monthly Raw Data in given timeframe.xlsx"
  unify_hfr("TZA", path, "~/WeeklyData/Output") }
}
