% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_hfr_data}
\alias{validate_hfr_data}
\title{Validate processed hfr data}
\usage{
validate_hfr_data(hfr_data, levels, orgs, ims, dates)
}
\arguments{
\item{hfr_data}{processed hfr data}

\item{levels}{datim org levels}

\item{orgs}{datim org hierarchy}

\item{ims}{datim mechanisms}

\item{dates}{hfr valid dates}
}
\value{
errors data frame
}
\description{
Validate processed hfr data
}
\examples{
\dontrun{
  validate_hfr_data(hfr_df, levels=org_levels, orgs=org_hierarchy, ims=mechanis, dates=hfr_dates)
}
}
