% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_import.R
\name{validate_import}
\alias{validate_import}
\title{Validation on import}
\usage{
validate_import(df)
}
\arguments{
\item{df}{df create during \code{hfr_import()}}
}
\description{
Validation on import
}
