% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_initial.R
\name{validate_initial}
\alias{validate_initial}
\title{Initial Validation on Submitted template}
\usage{
validate_initial(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Initial Validation on Submitted template
}
