% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_mechanism}
\alias{validate_mechanism}
\title{Validate mechanism code}
\usage{
validate_mechanism(mechanisms, ou, mcode)
}
\arguments{
\item{mechanisms}{df of mechs}

\item{mech_code}{mech code}
}
\value{
vector c(valid_im,  mech_name)
}
\description{
Validate mechanism code
}
\examples{
\dontrun{
  validate_mechanism(ims, 'Angola', 16172)
}

}
