% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_orgunit}
\alias{validate_orgunit}
\title{Validate org unit uid}
\usage{
validate_orgunit(df_orgs, ou, uid)
}
\arguments{
\item{df_orgs}{org hierarchy}

\item{ou}{operating unit}

\item{uid}{orgunituid}
}
\value{
valid as a vector c(valid_uid, valid_uid_ou)
}
\description{
Validate org unit uid
}
\examples{
\dontrun{
  validate_orgunit(df_orgs, 'Eswatini', 'g48XD8px8NN')
}
}
