% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_output.R
\name{validate_output}
\alias{validate_output}
\title{Validation Checks}
\usage{
validate_output(df)
}
\arguments{
\item{df}{HFR data framed created by \code{hfr_process_template()}}
}
\description{
Validation Checks
}
