% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_submission}
\alias{validate_submission}
\title{Validate submitted files}
\usage{
validate_submission(pfile, levels, ims)
}
\arguments{
\item{levels}{org levels}

\item{ims}{mechanisms df}

\item{sfile}{processed file}
}
\value{
vector c("fy", "hfr_pd", "iso3", "operatingunit", "mech_code", "mech_valid", "mech_name", "pdate", "name")
}
\description{
Validate submitted files
}
\examples{
\dontrun{
  validate_submissions("HFR_2020.99_XWH_100000_processed_20200101.csv")
}

}
