% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_submissions}
\alias{validate_submissions}
\title{Validate files}
\usage{
validate_submissions(pfolder, levels, ims, pattern = NULL)
}
\arguments{
\item{levels}{org levels}

\item{ims}{mechanisms df}

\item{pattern}{filename pattern}

\item{folder}{pfolder}
}
\value{
dataframe
}
\description{
Validate files
}
\examples{
\dontrun{
  validate_submissions(dir_hfr_pd205, pattern = "HFR_2020.05")
}

}
