% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{var_exists}
\alias{var_exists}
\title{Check if variable exist}
\usage{
var_exists(df, var)
}
\arguments{
\item{df}{data frame to check against}

\item{var}{quoted variable of interest}
}
\description{
Check if variable exist
}
