% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wavelength-package.R
\docType{package}
\name{Wavelength-package}
\alias{Wavelength}
\alias{Wavelength-package}
\title{Wavelength: Wavelength}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

USAID OHA Office. Munging of mission weekly HFR data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/Wavelength}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/Wavelength/issues}
}

}
\author{
\strong{Maintainer}: Aaron Chafetz \email{achafetz@usaid.gov}

}
\keyword{internal}
