% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_api.R
\name{extract_targets}
\alias{extract_targets}
\title{Extract Targets (DATIM API Call)}
\usage{
extract_targets(mechanism_id = NULL, ou_name = NULL, username,
  password, baseurl = "https://final.datim.org/")
}
\arguments{
\item{mechanism_id}{DATIM mechanism ID}

\item{ou_name}{Operating Unit name, if mechanism is not specified}

\item{username}{DATIM username}

\item{password}{DATIM password, recommend using \code{mypwd()}}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Extract Targets (DATIM API Call)
}
\examples{
\dontrun{
 #mechanism targets
 myuser <- "UserX"
 mech_x_targets <- extract_targets(mechanism_id = 00001, username = myuser, password = mypwd(myuser))
 #ou targets
 mech_x_targets <- extract_targets(ou_name = "Namibia", username = myuser, password = mypwd(myuser))
 }
}
