% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_date.R
\name{fix_date}
\alias{fix_date}
\title{Convert dates to date format}
\usage{
fix_date(df)
}
\arguments{
\item{df}{HFR data frame imported via \code{import_hfr()}}
}
\description{
Convert dates to date format
}
