% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_api.R
\name{gen_url}
\alias{gen_url}
\title{Generate a API URL}
\usage{
gen_url(mech_uid, ou_uid, org_lvl, org_type = "facility",
  type_hts = FALSE, curr_fy = 2019,
  baseurl = "https://final.datim.org/")
}
\arguments{
\item{mech_uid}{UID for a specific mechanism, recommend using \code{identify_mechs()}}

\item{ou_uid}{UID for the country, recommend using \code{identify_ouuids()}}

\item{org_lvl}{org hierarchy level, eg facility is level 7 in country X, recommend using \code{identify_levels()}}

\item{org_type}{organization type, either facility (default) or community}

\item{type_hts}{is the API call for HTS indicators, default = TRUE}

\item{curr_fy}{current fiscal year, default = 2019}

\item{baseurl}{API base url, default = https://final.datim.org/}
}
\description{
Generate a API URL
}
\examples{
\dontrun{
 #get mech UID
  mechuid <- identify_mechs(mechid = 00001, extract = "uid")
 #get OU UID
  ouuid <- identify_ouuids() \%>\% dplyr::filter(ou == "Ghana")
 #get facility level
  faclvl <- identify_levels("Ghana", "facility", username = myuser, password = mypwd())
 #gen url
  myurl <- gen_url(mechuid, ouuid, faclvl, org_type = facility, type_hts = FALSE) }
}
