% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{get_operatingunit}
\alias{get_operatingunit}
\title{Get operating unit name}
\usage{
get_operatingunit(orglevels, iso_code)
}
\arguments{
\item{orglevels}{df org levels}

\item{iso_code}{iso3 code}
}
\value{
operating unit
}
\description{
Get operating unit name
}
