% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_checks.R
\name{is_orgunituid_valid}
\alias{is_orgunituid_valid}
\title{Check if Operating Unit name is valid}
\usage{
is_orgunituid_valid(.data, df_orgs)
}
\arguments{
\item{df_orgs}{orgs}

\item{data}{df}
}
\value{
boolean
}
\description{
Check if Operating Unit name is valid
}
\examples{
\dontrun{
  data \%>\% is_orgunituid_valid(df_orgs)
}

}
