% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_map-data.R
\docType{data}
\name{iso_map}
\alias{iso_map}
\title{ISO Code Mapping Table}
\format{A data frame with each PEPFAR OU and associated ISO code.
\describe{
\item{operatingunit}{PEPFAR Operating Unit}
\item{countryname}{country under PEPFAR Operating Unit for regional programs}
\item{iso_ou}{3 letter ISO code for the Operating Unit}
\item{iso}{3 letter ISO code for the country}
}}
\source{
\url{https://github.com/ICPI/DIV/blob/master/Reference_Materials/2017.12.15_PEPFAR_ISOcodes.csv}
}
\usage{
data(iso_map)
}
\description{
Table that cross references the ISO Country codes to the
PEPFAR Operating Units and Countries. This table is used
during the export process for file naming. ISO codes were
originall pulled from Wikipedia and adapted for regional programs.
}
\keyword{datasets}
