% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lookups.R
\name{load_lookups}
\alias{load_lookups}
\title{Load DATIM Look up tables}
\usage{
load_lookups(datim_path = "./Data/", local = TRUE, user = NULL)
}
\arguments{
\item{datim_path}{HFR Data folder ID}

\item{local}{Read data from local directory?}

\item{user}{User email address}
}
\description{
Load DATIM Look up tables
}
\examples{
\dontrun{
 #load look up data
 load_lookups(datim_path = "./Data/datim")
 load_lookups(datim_path = "999#1aw####")
 }
}
