% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mech_pull.R
\name{mech_pull}
\alias{mech_pull}
\title{Pull Partner/Mechanism Info from DATIM}
\usage{
mech_pull()
}
\description{
Pull Partner/Mechanism Info from DATIM
}
\examples{
\dontrun{
#pull mechanism/partner information
df <- mech_pull() }
}
