% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datim_api.R
\name{periodize_targets}
\alias{periodize_targets}
\title{Adjust Annual Targets to work with HFR peirods}
\usage{
periodize_targets(df, quarters_complete)
}
\arguments{
\item{df}{data frame created by \code{extract_targets()}}

\item{quarters_complete}{MER quarters with data available}
}
\description{
Adjust Annual Targets to work with HFR peirods
}
\examples{
\dontrun{
 myuser <- "UserX"
 mech_x_targets <- extract_targets(00001, myuser, mypwd(myuser))
 mech_x_targets <- periodize_targets(mech_x_targets, 2) }
}
