% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_template.R
\name{process_template}
\alias{process_template}
\title{Import and Munge HFR Standard Templates}
\usage{
process_template(filepath)
}
\arguments{
\item{filepath}{filepath to sumbitted template}
}
\description{
Import and Munge HFR Standard Templates
}
