% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{report_submissions_errors}
\alias{report_submissions_errors}
\title{Report files validation}
\usage{
report_submissions_errors(df_files, mechanisms, export = FALSE)
}
\arguments{
\item{df_files}{df of filename validation}

\item{mechanisms}{mechs}
}
\value{
void
}
\description{
Report files validation
}
\examples{
\dontrun{
  report_file_errors(files, ims, processed)
}

}
