% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_date}
\alias{validate_date}
\title{Validate HFR PD Date}
\usage{
validate_date(df_pds, pdate, pd)
}
\arguments{
\item{df_pds}{hfr period dates}

\item{pdate}{period date}

\item{pd}{reporting period}
}
\value{
valid: True / False
}
\description{
Validate HFR PD Date
}
\examples{
\dontrun{
  validate_date(df_pds = valid_dates, '2020-01-27', 5)
}

}
