% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfr_dqa_utilities.R
\name{validate_submissions}
\alias{validate_submissions}
\title{Validate files}
\usage{
validate_submissions(levels, ims, pfolder, pattern = NULL)
}
\arguments{
\item{levels}{org levels}

\item{ims}{mechanisms df}

\item{folder}{pfolder}
}
\value{
dataframe
}
\description{
Validate files
}
